Error handling

+++ src/main/java/com/burrich/spring_running/address/AddressServiceImpl.java
@@ -3,7 +3,8 @@ package com.burrich.spring_running.address;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
+// TODO: error handling (create) => model validation
+// TODO: exist method (refactoring)
 
 @Service
 public class AddressServiceImpl implements AddressService {
@@ -17,8 +18,10 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
-    public Optional<Address> findById(Integer id) {
-        return addressRepository.findById(id);
+    public Address findById(Integer id) {
+        return addressRepository
+                .findById(id)
+                .orElseThrow(() -> new AddressNotFoundException(id));
     }
 
     @Override
@@ -29,7 +32,7 @@ public class AddressServiceImpl implements AddressService {
     @Override
     public Address update(Integer id, Address address) {
         if (!addressRepository.existsById(id)) {
-            return null;
+            throw new AddressNotFoundException(id);
         }
 
         address.setId(id);
@@ -38,6 +41,10 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void delete(Integer id) {
+        if (!addressRepository.existsById(id)) {
+            throw new AddressNotFoundException(id);
+        }
+
         addressRepository.deleteById(id);
     }
 }
